+++ portal/src/resources.ts
